+++ pyfakefs/tests/fake_filesystem_test.py
@@ -664,6 +664,15 @@ class FakeFilesystemUnitTest(TestCase):
         self.assertTrue(self.filesystem.exists(path))
         self.assertTrue(self.filesystem.exists(target_path))
 
+    def test_create_cross_device_link(self):
+        self.filesystem.add_mount_point("/mnt/dir1")
+        source = "/mnt/dir1/myfile.txt"
+        dest = "/mnt/dir2/myotherfile.txt"
+        self.filesystem.create_file(source)
+
+        with self.assertRaisesRegex(OSError, "Invalid cross-device link"):
+            self.filesystem.create_link(source, dest)
+
     def test_resolve_object(self):
         target_path = "dir/target"
         target_contents = "0123456789ABCDEF"